python-samples/gcp-key-management/requirements.txt
google-cloud-resource-manager==1.10.1
google-cloud-kms==2.17.0